fix issues
